requirements.txt
asgiref==3.7.2
cfgv==3.4.0
distlib==0.3.8
Django==5.0
django-filter==23.5
djangorestframework==3.14.0
drf-yasg==1.21.7
easy-thumbnails==2.8.5
filelock==3.13.1
flake8==6.1.0
identify==2.5.33
inflection==0.5.1
isort==5.13.2
mccabe==0.7.0
nodeenv==1.8.0
packaging==23.2
pillow==10.2.0
platformdirs==4.1.0
pre-commit==3.6.0
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
virtualenv==20.25.0
